R For Data Science (2e) Exercises
RMD workups for each section of the R for Data Science (2e) text. I worked through the given Exercises throughout each chapter. Given the size of the work required, I've broken up each section of the book into seperate RMDs. This first upload covers work completed so far: Whole Game & Visualize. I plan to complete the rest of the book's Exercises and upload them here. (NOTE: this is not a required assignment for class, although we do read the text. This is extra work I performed to ensure my understanding of the text.)